+++ src/components/tabs/index.js
@@ -1,1 +1,0 @@
-export { Tabs, Tab } from "./tabs.component";